refactor: fix path and names
